+++ src/components/Project/Resource/DatasetPage.tsx
@@ -25,6 +25,7 @@ import DatasetIcon from "@mui/icons-material/FormatListBulleted";
 import { AddFromServiceDialog } from "@app/components/Project/Resource/AddFromServiceDialog";
 import { CreateDatasetDialog } from "@app/components/Project/Resource/CreateDatasetDialog";
 import TableDataModal from "@app/components/TableDataModal";
+import { IncoreDialog } from "@app/components/IncoreDialog";
 
 const DatasetPage = (): JSX.Element => {
     const { id } = useParams(); // Get projectId from the URL path
@@ -144,8 +145,35 @@ const DatasetPage = (): JSX.Element => {
     const onCreateDataset = () => {
         setOpenCreateDatasetDialog(true);
     };
+
+    // batch delete
+    const [selectedDatasets, setSelectedDatasets] = useState<Dataset[]>([]);
+    const [openBatchDeleteDialog, setOpenBatchDeleteDialog] = useState(false);
+    const handleBatchDelete = async () => {
+        if (project?.id && selectedDatasets.length > 0) {
+            await appDispatch(
+                deleteProjectDatasets({
+                    projectId: project.id,
+                    datasetIds: selectedDatasets.map((w) => w.id)
+                })
+            );
+            setSelectedDatasets([]);
+        }
+        setOpenBatchDeleteDialog(false);
+    };
+
     return (
         <Container sx={{ display: "flex", flexDirection: "column", height: "100vh" }} maxWidth="xl">
+            <IncoreDialog
+                open={openBatchDeleteDialog}
+                onClose={() => {
+                    setOpenBatchDeleteDialog(false);
+                }}
+                onAction={handleBatchDelete}
+                message="Are you sure you want to delete the selected items? This action cannot be undone."
+                dialogTitle="Confirm Deletion"
+                actionButtonName="Batch Delete"
+            />
             <Box sx={{ flexShrink: 0 }} mt={5}>
                 {!project ? (
                     <Typography>Loading...</Typography>
@@ -175,6 +203,8 @@ const DatasetPage = (): JSX.Element => {
                                     createLabel="Add from Service"
                                     addtionalCreateLabel="Upload Dataset"
                                     additionalCreateClick={onCreateDataset}
+                                    selectedItemsCount={selectedDatasets.length}
+                                    onBatchDeleteClick={() => setOpenBatchDeleteDialog(true)}
                                 />
                                 <AddFromServiceDialog
                                     projectId={project.id}
@@ -203,6 +233,8 @@ const DatasetPage = (): JSX.Element => {
                                             setOpenTableDataModal(true);
                                             setSelectedDataset(dataset);
                                         }}
+                                        onSelectionChange={(selected) => setSelectedDatasets(selected as Dataset[])}
+                                        selectedItems={selectedDatasets}
                                     />
                                 ) : (
                                     <ResourceCards
@@ -215,6 +247,8 @@ const DatasetPage = (): JSX.Element => {
                                             setOpenTableDataModal(true);
                                             setSelectedDataset(dataset);
                                         }}
+                                        onSelectionChange={(selected) => setSelectedDatasets(selected as Dataset[])}
+                                        selectedItems={selectedDatasets}
                                     />
                                 )}
                                 <Box mt={4} display="flex" justifyContent="center">